+++ src/regression/suite/config/invalid-test-case.error.ts
@@ -3,7 +3,7 @@ import {TestSuiteConfig} from './test-suite-config';
 export class InvalidTestCaseError extends Error {
 
     constructor(public testCase: TestSuiteConfig, message: string) {
-        super(`Invalid test case '${testCase.testCase}': ${message}`);
+        super(`Invalid test case '${testCase.testSuite}': ${message}`);
     }